refactor(listener): extract settings commands
